Add output 'vacuumsucc', modify break behavior Break function will now kill the yellow wool block item that was dropped by lucky blocks.
New output:
Name: Vaccuumsucc
Type: Negative
Is scripted?: True
Desc: Creates a 'vacuum' entity at it's break location, and pulls in
all entities (with the exception of item frames) in a distance of 25
blocks to itself. Players are pulled in as well, but are only pulled in
until they're 5 blocks or less from the entity. Despawns after 10
seconds.